http_server: fix vibe-coded bug
`mk_vhost_handler` expects a regex, not a glob-star kind of match.

To fix, we actually move the input name to the end of the URL. The regex
library used here does not support negative character-class matches, and
so it would be pretty difficult to embed the input name in the middle of
the URL (as one would expect from REST-like conventions) without
parsing manually character-by-character. Instead, we just use a `.*` and
try to find the end of the string (which I'm probably not doing
correctly).